requirements.txt
joblib==1.5.2
requests==2.32.5
polars==1.34.0
torch==2.8.0
torchvision==0.23.0
xgboost==3.0.5
lightgbm==4.6.0
tensorflow_cpu==2.20.0
scikit-learn==1.7
pandas==2.3.3
pyarrow==21.0.0
